RIA-6040 fix LR confirmation bug
